% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Sylamer_Tools.R
\name{create_sylamer_script}
\alias{create_sylamer_script}
\title{Generate shell script to run Sylamer in one or multiple sorted lists of genes}
\usage{
create_sylamer_script(
  utrFile = "dm3_flybase_3utrs.fa",
  kSylamer = 6,
  kMarkov = 4,
  winSize = 200,
  sylamer = "/opt/linux64bin/sylamer",
  sylOutput = "shRNA_sylamer.output.tab",
  imageOut = "shRNA_image.output.pdf",
  DEsorted = c("ForSylamer.txt", "ForSylamer_2.txt"),
  extras = "-a 10 -aa 5 --funny-ok",
  overwrite = T,
  sylamer_comand = "sylamer_command.sh"
)
}
\arguments{
\item{utrFile}{File with the UTR sequences. Defauls is "dm3_flybase_3utrs.fa"}

\item{kSylamer}{Oligo letter length or Word size, has to be the same length as those in the "words" file. Default is 6}

\item{kMarkov}{Size of a smaller word to be used for correcting composition biases. Default is 4}

\item{winSize}{How many sequences should be added in each consecutive window. Default is 200}

\item{sylamer}{Sylamer executable path. Default is "/opt/linux64bin/sylamer"}

\item{sylOutput}{Name of output file to write table in. Default is "shRNA_sylamer.output.tab"}

\item{imageOut}{Name of output file to write plot in. Default is "shRNA_image.output.pdf"}

\item{DEsorted}{Name of one or multiple list of sorted genes to input in sylamen. Default is c("ForSylamer.txt","ForSylamer_2.txt")}

\item{extras}{Any extra parametere to pass into Sylamer. Default is "-a 10 -aa 5 --funny-ok"}

\item{overwrite}{Boolean. If True, the script will overwrite any existing script with the same name. Default is T}

\item{sylamer_comand}{Name of the file to write the command into "sylamer_command.sh"}
}
\value{
a file ready to run "sh sylamer_command.sh"
}
\description{
Generate shell script to run Sylamer in one or multiple sorted lists of genes
}
\examples{

#create_sylamer_script(utrFile="dm3_flybase_3utrs.fa",kSylamer=6,kMarkov =4,winSize=200,sylamer="/opt/linux64bin/sylamer",sylOutput="shRNA_sylamer.output.tab",imageOut="shRNA_image.output.pdf",DEsorted=c("ForSylamer.txt","ForSylamer_2.txt"), extras = "-a 10 -aa 5 --funny-ok", overwrite=T,sylamer_comand="sylamer_command.sh")


}
