% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Sylamer_Tools.R
\name{sylamerplots}
\alias{sylamerplots}
\title{Plot the output of Sylamer}
\usage{
sylamerplots(
  oligosChosen = c("TGTAAA", "GATGCT"),
  sylamer.out.tab = "shRNA_sylamer.output.tab",
  topOligos = 0,
  lowOligos = 0,
  outName = "sylamerplot",
  greylines = F,
  kSylamer = 6,
  kMarkov = 4,
  winSize = 200,
  plot = T,
  free_ylim = F
)
}
\arguments{
\item{oligosChosen}{Sequences to highlight (Any particular words that you want highlighted) Defaults is =c("TGTAAA","GATGCT")}

\item{sylamer.out.tab}{Name of sylamer output to read from. Dafault is "shRNA_sylamer.output.tab"}

\item{topOligos}{Number of top oligos to plot. How many best words should I highlight in the image? Defauls is 0}

\item{lowOligos}{Number of low oligos to plot. Defauls is 0}

\item{outName}{name for the plot output. Default is "sylamerplot"}

\item{greylines}{Boolean, To plot or not greay background oligos. Default is F}

\item{kSylamer}{Oligo letter length or Word size, has to be the same length as those in the "words" file. Default is 6}

\item{kMarkov}{Size of a smaller word to be used for correcting composition biases. Default is 4}

\item{free_ylim}{Boolean, if TRUE then y lim will be the min and max values from Sylamr output. If FALSE, then ylim=(-10,10)}
}
\value{
a file ready to run "sh sylamer_command.sh"
}
\description{
Plot the output of Sylamer
}
\examples{

#create_sylamer_script(utrFile="dm3_flybase_3utrs.fa",kSylamer=6,kMarkov =4,winSize=200,sylamer="/opt/linux64bin/sylamer",sylOutput="shRNA_sylamer.output.tab",imageOut="shRNA_image.output.pdf",DEsorted=c("ForSylamer.txt","ForSylamer_2.txt"), extras = "-a 10 -aa 5 --funny-ok", overwrite=T,sylamer_comand="sylamer_command.sh")


}
